[linux] move debug-trap.h and winuser.rh to libs/common/include
