+++ src/libs/common/include/debug-trap.h
